+++ BirthdayBot/Cogs/Registration.py
@@ -3,30 +3,16 @@ import discord
 from discord.ext import commands
 import datetime
 from datetime import datetime
-from discord.ui import Button, View
-from BirthdayBot.Utils import session_scope, logger
+from BirthdayBot.Utils import session_scope, logger, addUserToDB
 from BirthdayBot.Models import DiscordUser
-from sqlalchemy.exc import SQLAlchemyError
-from BirthdayBot.Cogs.UserAgeInfo import UserAgeInfo
-from BirthdayBot.Cogs import Help
+from BirthdayBot.Views import RegistrationButtons, ExistingUserButtons
 
-# NOTES:
-##### MAKE IT SO THE MESSAGE SENDER SEE CONFIRMATION + PLEASE TRY AGAIN + EVERYTHING ELSE
-#####SEND BIRTHDAY MESSAGE LOGIC FOR USERS IN OTHER DISCORDS
-#####FOR every guild.id check the guild for names in the database if the name is in the database AND in the guild.id then they dont need to re-enter bday for that guild "You already entered your birthday in another guild"
 class Registration(commands.Cog):
     """Class Dedicated to housing all commands related to registration"""
-
     def __init__(self, bot):
         self.bot = bot
 
-    ##########################################################################
-    async def test(self, ctx):
-        await ctx.send("test")
-
-    ##########################################################################
     """ ---- COMMANDS ---- """
-
     @commands.hybrid_command(
         name="register",
         description="Prompts the user with a message to register their birthday.",
@@ -84,7 +70,7 @@ class Registration(commands.Cog):
             await ctx.send("Timed out")
         elif view.userConfirmation:
             try:
-                self.writeUserToDB(
+                addUserToDB(
                     username=msg.author,
                     birthday=msg.content,
                     discord_id=msg.author.id,
@@ -151,7 +137,7 @@ class Registration(commands.Cog):
                 outerLoop = False
             elif view.userConfirmation and update == False:
                 try:
-                    self.writeUserToDB(
+                    addUserToDB(
                         username=msg.author,
                         birthday=msg.content,
                         discord_id=msg.author.id,
@@ -281,82 +267,6 @@ class Registration(commands.Cog):
         )
         await ctx.send(embed=embed)
 
-    @staticmethod
-    def writeUserToDB(username: str, birthday: str, discord_id: str, guild: int):
-        # session_scope will raise an exception if invalid, use this with try/except
-        with session_scope() as s:
-            user = DiscordUser(
-                username=str(username),
-                Birthday=birthday,
-                discord_ID=discord_id,
-                guild=guild,
-            )
-            s.add(user)
-
 
 async def setup(bot):
     await bot.add_cog(Registration(bot))
-
-
-class RegistrationButtons(discord.ui.View):
-    def __init__(self, *, timeout=180, author):
-        super().__init__(timeout=timeout)
-        self.userConfirmation = None
-        self.author = author
-
-    @discord.ui.button(label="Yes! 👍", style=discord.ButtonStyle.green)  # or .success
-    async def yes(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message(
-            "Confirming..."
-        )  # Ephermal = True if we only want user to see, tbd
-        self.userConfirmation = True
-        self.stop()
-
-    @discord.ui.button(label="No! 👎", style=discord.ButtonStyle.red)  # or .danger
-    async def no(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message("Please try again... (mm/dd/yyyy)")
-        self.userConfirmation = False
-        self.stop()
-
-    async def interaction_check(self, inter: discord.MessageInteraction) -> bool:
-        if inter.user != self.author:
-            await inter.response.send_message(
-                content="You don't have permission to press this button.",
-                ephemeral=True,
-            )
-            return False
-        return True
-
-
-class ExistingUserButtons(discord.ui.View):
-    def __init__(self, *, timeout=180, author, existing_user: DiscordUser):
-        super().__init__(timeout=timeout)
-        self.userConfirmation = None
-        self.author = author
-        self.existing_user_bday = existing_user.Birthday
-
-    @discord.ui.button(label="Yes!", style=discord.ButtonStyle.green)  # or .success
-    async def yes(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message(
-            "Please Provide a new Birthday...(mm/dd/yyyy)"
-        )  # Ephermal = True if we only want user to see, tbd
-        self.userConfirmation = True
-        self.stop()
-
-    @discord.ui.button(label="No!", style=discord.ButtonStyle.red)  # or .danger
-    async def no(self, interaction: discord.Interaction, button: discord.ui.Button):
-        daysAway = UserAgeInfo.daysAway(birthdate=self.existing_user_bday)
-        await interaction.response.send_message(
-            "Sounds good! Only {} Days from your birthday!".format(daysAway)
-        )
-        self.userConfirmation = False
-        self.stop()
-
-    async def interaction_check(self, inter: discord.MessageInteraction) -> bool:
-        if inter.user != self.author:
-            await inter.response.send_message(
-                content="You don't have permission to press this button.",
-                ephemeral=True,
-            )
-            return False
-        return True
